requirements.txt
pyannote.audio
pydub
flask
datasets
python-dotenv
requests